requirements.txt
streamlit
Pillow
watchdog


# streamlit run /Users/leonard/Desktop/Repositories/Leonard.github.io/app.py